requirements.txt
uvicorn[standard]>=0.20.0
fastapi>=0.95.0
sqlalchemy>=2.0.0
geoalchemy2>=0.13.0
psycopg2-binary>=2.9.3
